Proposed change to the way "i :=" works
http://johnearnest.github.io/Octo/index.html?gist=fb284783560e65edb8f9 minimal test-case
After setting "i" to point somewhere that requires using a long memory reference, when you then set it to point somewhere else without using the special long memory reference code, it won't change the leftmost byte of its address. This causes it to not actually point where you requested it to.

I'm not sure if this is intended behaviour or not but I ran into it when making civ8n and it confused me for a bit. It seems intended based on the code but I thought I'd try this whole forking/pull request thing anyway.